[ project ] organized folder and files
