--- 172/project/Entrega1LaTeX/sigproc.bib
@@ -1,29 +1,0 @@
-@BOOK{Cormen:Algorithms,
-	AUTHOR = "Thomas H. Cormen and Charles E. Leiserson and Ronald L. Rivest and Clifford Stein",
-	TITLE = "Introduction to algorithms",
-	PUBLISHER = "The MIT Press",
-	ADDRESS = "Cambrigde, Massachusetts",
-	YEAR = "2001"	}
-
-
-@Misc{Black:RBTree,
-  title =	 {Dictionary of algorithms and data structures: red-black tree},
-  author =	 {Paul E. Black},
-	year =	 "2015",
-  note = {Retrieved August 7, 2017 from \url{https://xlinux.nist.gov/dads/HTML/redblack.html}}
-}
-
-
-	@online{Tuto:Data,
-  author = "TutorialsPoint",
-  year = "2016",
-  title = "Data structures and algorithms",
-  note = {Retrieved August 7, 2017 from \url{http://www.tutorialspoint.com/data_structures_algorithms/data_structures_algorithms_tutorial.pdf}},
-}
-
-@BOOK{Mehta:Handbook,
-  title={Handbook of data structures and applications},
-  author={Mehta, Dinesh P and Sahni, Sartaj},
-  year={2004},
-  publisher={CRC Press}
-}
